requirements.txt
asgiref==3.8.1
Django==5.2.1
djangorestframework==3.16.0
dotenv==0.9.9
pillow==11.2.1
psycopg==3.2.9
python-dotenv==1.1.0
sqlparse==0.5.3
tzdata==2025.2